Cythonize the call loop
Cythonize `pluggy.callers._multicall` which gives around a 2x speedup
according to the benchmark suite. Transform the `pluggy.callers`
module in a new package and move utils and the legacy call loop into
separate modules.
